worker/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.0
pymongo==4.5.0
apscheduler==3.10.4